requirements.txt
asgiref==3.5.2
Django==4.0.5
django-filter==22.1
djangorestframework==3.13.1
importlib-metadata==4.11.4
Markdown==3.3.7
PyJWT==1.7.1
pytz==2022.1
sqlparse==0.4.2
zipp==3.8.0
